Added Checking and Savings classes
